+++ src/main/java/view/InputView.java
@@ -1,7 +1,7 @@
 package view;
 
-import domain.Player;
 import domain.Team;
+import domain.player.Player;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
